Move `VisitMember` implementations to base class
This introduces two small, benign changes:

 * `Mock`'s visitor now has an unused field `isFirst`.
 * LINQ's visitor now allows compiler-generated field accesses.
