+++ src/Moq/Linq/FluentMockVisitor.cs
@@ -14,15 +14,6 @@ namespace Moq.Linq
 {
 	internal class FluentMockVisitor : FluentMockVisitorBase
 	{
-		/// <summary>
-		/// The first method call or member access will be the 
-		/// last segment of the expression (depth-first traversal), 
-		/// which is the one we have to Setup rather than FluentMock.
-		/// And the last one is the one we have to Mock.Get rather 
-		/// than FluentMock.
-		/// </summary>
-		private bool isFirst = true;
-
 		protected override Expression VisitParameter(ParameterExpression node)
 		{
 			if (node == null)
@@ -52,45 +43,8 @@ namespace Moq.Linq
 			return TranslateFluent(node.Object.Type, node.Method.ReturnType, targetMethod, Visit(node.Object), lambdaParam, lambdaBody);
 		}
 
-		protected override Expression VisitMember(MemberExpression node)
-		{
-			if (node == null)
-			{
-				return null;
-			}
-
-			// If member is not mock-able, actually, including being a sealed class, etc.?
-			if (node.Member is FieldInfo)
-			{
-				throw new NotSupportedException();
-			}
-
-			// Translate differently member accesses over transparent
-			// compiler-generated types as they are typically the 
-			// anonymous types generated to build up the query expressions.
-			if (node.Expression.NodeType == ExpressionType.Parameter &&
-				node.Expression.Type.GetTypeInfo().IsDefined(typeof(CompilerGeneratedAttribute), false))
-			{
-				var memberType = ((PropertyInfo)node.Member).PropertyType;
-
-				// Generate a Mock.Get over the entire member access rather.
-				// <anonymous_type>.foo => Mock.Get(<anonymous_type>.foo)
-				return Expression.Call(null, Mock.GetMethod.MakeGenericMethod(memberType), node);
-			}
-
-			var lambdaParam = Expression.Parameter(node.Expression.Type, "mock");
-			Expression lambdaBody = Expression.MakeMemberAccess(lambdaParam, node.Member);
-			var targetMethod = GetTargetMethod(node.Expression.Type, ((PropertyInfo)node.Member).PropertyType);
-			if (isFirst)
-			{
-				isFirst = false;
-			}
-
-			return TranslateFluent(node.Expression.Type, ((PropertyInfo)node.Member).PropertyType, targetMethod, Visit(node.Expression), lambdaParam, lambdaBody);
-		}
-
 		// Args like: string IFoo (mock => mock.Value)
-		private static Expression TranslateFluent(
+		protected override Expression TranslateFluent(
 			Type objectType,
 			Type returnType,
 			MethodInfo targetMethod,
@@ -109,7 +63,7 @@ namespace Moq.Linq
 			return Expression.Call(instance, targetMethod, Expression.Lambda(funcType, lambdaBody, lambdaParam));
 		}
 
-		private MethodInfo GetTargetMethod(Type objectType, Type returnType)
+		protected override MethodInfo GetTargetMethod(Type objectType, Type returnType)
 		{
 			MethodInfo targetMethod;
 			// dte.Solution =>
